переносит валидацию в отдельный блок
